fix: clippy 1.96 lints the dev CI toolchain catches (collapsible_if, unnecessary_cast)
CI runs clippy on a newer toolchain than the local one — collapse the
compose_plan network-check into a let-chain and drop a no-op cast in the
seed_files test decoder. No behavior change (compose 17 + seed 18 tests
green).

+++ crates/snapshot/src/docker_import/seed_files.rs
@@ -634,7 +634,7 @@ mod tests {
                 n = (n << 6) | val(c).unwrap();
                 bits += 6;
             }
-            let mut shift = bits as i32 - 8;
+            let mut shift = bits - 8;
             while shift >= 0 {
                 out.push(((n >> shift as u32) & 0xff) as u8);
                 shift -= 8;
